// front/src/providers/api-client.generated/model/sendCandidateByEmailRequest.ts
/**
 * Morgan And Mallet CRM
 * Morgan And Mallet CRM API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CandidateFileDto } from './candidateFileDto';
import { CandidateResumeOptions } from './candidateResumeOptions';


export interface SendCandidateByEmailRequest { 
    /**
     * The start of the request
     */
    start?: number;
    /**
     * The length of the request
     */
    length?: number;
    /**
     * order by field
     */
    orderby?: string;
    /**
     * order direction (asc | desc)
     */
    order?: string;
    /**
     * Search
     */
    search?: string;
    to: string;
    subject: string;
    body: string;
    customerName: string;
    mode: string;
    candidatePublicLink?: string;
    candidateId?: string;
    candidateFiles?: Array<CandidateFileDto>;
    candidatesIds?: string;
    candidateResumeOptions?: Array<CandidateResumeOptions>;
}